feat: fix reading of config in oas2ts
