+++ example/json-schema-to-typescript-config.json
